requirements.txt
pytesseract 
pdf2image 
pillow
opencv-python
fitz
spacy
python-dotenv
groq
streamlit